--- Parser/Prueba0.txt
@@ -1,18 +1,0 @@
-int variable1 variable1
-+
-auto
-==
-"SñlksfR
-ET}"
-//hi
-/* Hola
-HGPOGF
-*/
-'\x00'
-'\t'
-ñ
-ñ
-ññ
-'
-'
-<<
