--- Parser/Prueba2.txt
@@ -1,9 +1,0 @@
-
-//Todos son identificadores. La idea es que distinga case sensitive 
-// y cuente las ocurrencias por linea
-
-X X Y
-Y
-S  S x
-S s
-X Y
