--- Parser/Prueba4.txt
@@ -1,38 +1,0 @@
-#include <stdio.h>
-#include <string.h>
-#define MSG “Introduce tira de caracteres: ”
-
-void compara (char a[], char b[]) {
-    int i=strcmp(a,b);
-    if (!i)
-        printf("iguales"); /* falso <-> i=0 */
-    else if (i<0)
-        printf("%s > %s",a,b);
-    else
-        printf("%s < %s",a,b);
-}
-void concatena (char a[], char b[]) {
-    if (strlen(a)+strlen(b)<80) {
-        strcat (a,b);
-        printf ("%s",a);
-        }
-    }
-}
-void copia (char a[], char b[]) {
-    strcpy(a,b);
-    printf ("%s y %s",a,b);
-}
-void longitud (char a[], char b[]) {
-    printf("%s = longitud %d: %s = longitud %d",a,strlen(a),b,strlen(b));
-}
-int main() {
-    char x[80], y[80]; /* vectores */
-    printf (MSG);
-    scanf ("%s",x);
-    printf (MSG);
-    scanf ("%s",y);
-    longitud(x,y);
-    compara(x,y);
-    concatena(x,y);
-    copia(x,y);
-}
