--- Parser/src/main/java/Main.java
@@ -1,12 +1,0 @@
-
-public class Main {
-
-    private static String filePath = "integral.c";
-
-    public static void main(String[] args) {
-        TokenTable tokenTable = new TokenTable();
-        tokenTable.scanFile(filePath);
-        tokenTable.printTokenTable();
-        tokenTable.printErrorTable();
-    }
-}
