--- Parser/src/main/java/Tokens.java
@@ -1,88 +1,0 @@
-
-public enum Tokens {
-    AUTO, 
-    BREAK,
-    CASE,
-    CHAR,
-    CONST,
-    CONTINUE,
-    DEFAULT, 
-    DO, 
-    DOUBLE, 
-    ELSE,
-    ENUM,
-    EXTERN,
-    FLOAT,
-    FOR,
-    GOTO,
-    IF,
-    INT,
-    LONG,
-    REGISTER,
-    RETURN,
-    SHORT,
-    SIGNED,
-    SIZEOF,
-    STATIC,
-    STRUCT,
-    SWITCH,
-    TYPEDEF,
-    UNION,
-    UNSIGNED,
-    VOID,
-    VOLATILE,
-    WHILE,
-    DEC_LIT,
-    OCTAL_LIT,
-    HEX_LIT,
-    DOUBLE_LIT,
-    STRING_LIT,
-    CHAR_LIT,
-    _IDENTIFIER,
-    COMMA,
-    SEMICOLON,
-    INCREMENT,
-    DECREMENT,
-    EQUAL_TO,
-    GRATER_EQUAL,
-    GRATER,
-    TERNARY,
-    LESS_EQUAL,
-    LESS,
-    NOT_EQUAL,
-    OR, 
-    AND,
-    NOT, 
-    EQUAL,
-    PLUS,
-    MINUS,
-    MULT,
-    DIV,
-    MODULE,
-    RBO,
-    RBC,
-    SBO,
-    SBC,
-    CBO,
-    CBC,
-    COLON,
-    DOT,
-    PLUS_ASSIGNMENT,
-    MINUS_ASSIGNMENT,
-    MULT_ASSIGNMENT,
-    DIV_ASSIGNMENT,
-    BIT_AND,
-    XOR,
-    BIT_OR,
-    RIGHT_SHIFT,
-    LEFT_SHIFT,
-    COMPLEMENT,
-    MOD_ASSIGNMENT,
-    AND_ASSIGNMENT,
-    XOR_ASSIGNMENT,
-    OR_ASSIGNMENT,
-    LEFT_SHIFT_ASSIGNMENT,
-    RIGHT_SHIFT_ASSIGNMENT,
-    OP,
-    ERROR
-}
