+++ cantidadLineas.c
