+++ literales.c
