+++ reservadas_operadores.c
